requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==20.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-js-asset==2.0.0
django-mptt==0.14.0
docker==6.0.1
docker-composer==0.8.5
exceptiongroup==1.0.4
factory-boy==3.2.1
Faker==15.3.4
h11==0.14.0
idna==3.4
iniconfig==1.1.1
loguru==0.5.3
outcome==1.2.0
packaging==22.0
Pillow==9.3.0
pkg_resources==0.0.0
pluggy==1.0.0
psycopg2-binary==2.9.2
psycopg2-pool==1.1
PySocks==1.7.1
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
webdriver-manager==3.8.5
websocket-client==1.4.2
wsproto==1.2.0
